#425: Split ItemEditorReducerTests into multiple files
